feat(core): extract video frame sampling into parity package
Move the pure ffmpeg seek+capture core of video frame extraction
(computeSeekTimestamps, extractFrame, extractFrames, extractFramesAt)
from apps/api/src/face/video-frame-extraction.service.ts into a new
packages/enrichment-compute/src/video/ subpath, verbatim, so a
distributed worker node can sample frames identically to the server —
groundwork for video_face_detection node support (distributed-nodes
spec §7 parity contract) and immediately usable by a node's
social-media-detection Tier-2 OCR pass.

Per-frame extraction failures now route through the package's
computeLog seam (wired to NestJS Logger once at
image-orientation.util.ts import time) instead of a service-local
Logger — same observable warning, no assertions in the existing spec
depend on the logger.

apps/api/src/face/video-frame-extraction.service.ts becomes a thin
delegating adapter; all 19 pre-existing
video-frame-extraction.service.spec.ts cases pass unmodified, proving
jest.mock('fluent-ffmpeg') still intercepts the package's nodeRequire
correctly through the delegation.

Adds the packages/enrichment-compute/video/ node10-resolution
back-compat stub package.json (mirrors face/ocr/metadata/dto/social)
since apps/api compiles with moduleResolution:"node".
